% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_with_chip_genes}
\alias{annotate_with_chip_genes}
\title{Annotate top hits with CHIP gene information}
\usage{
annotate_with_chip_genes(top_hits)
}
\arguments{
\item{top_hits}{A data frame containing the top hits.}
}
\value{
A data frame with the CHIP gene information.
}
\description{
Annotate top hits with CHIP gene information
}
