% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_with_immunoglobulin}
\alias{annotate_with_immunoglobulin}
\title{Annotate data with immunoglobulin gene information}
\usage{
annotate_with_immunoglobulin(x, ...)
}
\arguments{
\item{x}{A data frame/tibble or GWASFormatter object containing variant data.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Annotate data with immunoglobulin gene information
}
