% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ash.r
\name{append_ashr_results}
\alias{append_ashr_results}
\title{Append lfsr results from ashr to a data frame}
\usage{
append_ashr_results(x, ...)
}
\arguments{
\item{x}{A data frame containing GWAS summary statistics with columns "BETA" and "SE".}

\item{...}{Additional arguments passed to the method.}
}
\value{
A data frame with additional columns for lfsr, qvalue, pm, and psd.
}
\description{
This function appends lfsr results from the ashr package to a data frame containing GWAS summary statistics.
}
