% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_sample}
\alias{compute_sample}
\title{Take a random sample of rows from the summary stats table}
\usage{
compute_sample(g, rows = 100L)
}
\arguments{
\item{g}{A GWASFormatter object.}

\item{rows}{The number of rows to sample. Default is 100.}
}
\description{
Take a random sample of rows from the summary stats table
}
