% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_connect}
\alias{db_connect}
\title{Connect to DuckDB}
\usage{
db_connect(read_only = FALSE)
}
\arguments{
\item{read_only}{Logical. If TRUE, the connection will be read-only.}
}
\value{
A DuckDB connection object.
}
\description{
Connect to DuckDB
}
