% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_variants.R
\name{exclude_difficult_regions.GWASFormatter}
\alias{exclude_difficult_regions.GWASFormatter}
\title{Exclude difficult regions}
\usage{
\method{exclude_difficult_regions}{GWASFormatter}(
  x,
  beds_exclude = c("hg19diff", "UCSC_unusual", "GRC_exclusions"),
  active_table = "filtered_variants",
  ...
)
}
\arguments{
\item{beds_exclude}{A character vector of bed files to exclude. Options are "hg19diff", "UCSC_unusual", "GRC_exclusions", and "GIAB_difficult_regions".}

\item{active_table}{The name of the active table to exclude difficult regions from. Options are "filtered_variants" and "summary_stats".}
}
\description{
Exclude difficult regions
}
