% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_variants.R
\name{exclude_difficult_regions}
\alias{exclude_difficult_regions}
\title{Exclude difficult regions}
\usage{
exclude_difficult_regions(x, ...)
}
\arguments{
\item{x}{A GWASFormatter object.}

\item{...}{Additional arguments passed to the method.}
}
\description{
Exclude difficult regions from a GWASFormatter object.
}
