% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_variants.R
\name{filter_variants}
\alias{filter_variants}
\title{Filter variants}
\usage{
filter_variants(x, subset = NULL, exclude = NULL, ...)
}
\arguments{
\item{x}{A GWASFormatter object.}

\item{subset}{A file path to a whitelist of variants in Parquet format. Only variants in this file will be kept.}
}
\description{
Filter variants in a GWASFormatter object based on a whitelist or blacklist of variants.
}
