% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{format_variants_for_vep}
\alias{format_variants_for_vep}
\title{Format variants for VEP API input}
\usage{
format_variants_for_vep(variants)
}
\arguments{
\item{variants}{Input variants in format "chr_pos_ref_alt"}
}
\value{
Character vector of VCF-formatted variant strings
}
\description{
Format variants for VEP API input
}
\keyword{internal}
