% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{format_vep_results}
\alias{format_vep_results}
\title{Format VEP API results into a tidy data.frame}
\usage{
format_vep_results(vep_results, flag_pick = FALSE)
}
\arguments{
\item{vep_results}{Raw results from VEP API (can be data.frame or list)}

\item{flag_pick}{Logical. If TRUE, filter for transcripts with PICK flag}
}
\value{
Formatted data.frame
}
\description{
Format VEP API results into a tidy data.frame
}
\keyword{internal}
