% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locuszoom.R
\name{get_encode_biosamples}
\alias{get_encode_biosamples}
\title{Get available ENCODE SCREEN biosamples for a given assembly}
\usage{
get_encode_biosamples(assembly = "grch38")
}
\arguments{
\item{assembly}{Genome assembly ("grch38" or "mm10")}
}
\value{
Data frame with biosample metadata including cell type names
}
\description{
Get available ENCODE SCREEN biosamples for a given assembly
}
