% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwasplot-package.R
\docType{package}
\name{gwasplot-package}
\alias{gwasplot}
\alias{gwasplot-package}
\title{gwasplot: Publication-Ready and Interactive Plots for GWAS Results}
\description{
A ggplot2-based toolkit for visualising genome-wide association study (GWAS) results. Provides linear Manhattan plots, per-chromosome plots and circular (CMplot-style) plots with support for multiple concentric trait rings, easy highlighting and labelling of top markers, interactive versions via 'ggiraph', and companion tables of top hits via 'gt'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/loukesio/gwasplot}
  \item Report bugs at \url{https://github.com/loukesio/gwasplot/issues}
}

}
\author{
\strong{Maintainer}: loukesio \email{loukesio@gmail.com}

}
\keyword{internal}
