% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{qqplot_save}
\alias{qqplot_save}
\title{Save a QQ plot of GWAS p-values}
\usage{
qqplot_save(gwas, output_prefix, width = 5, height = 5, dpi = 300)
}
\arguments{
\item{gwas}{A gwas object containing the p-values to plot.}

\item{output_prefix}{The prefix for the output file name.}

\item{width}{Width of the output figure in inches.}

\item{height}{Height of the output figure in inches.}

\item{dpi}{DPI of the output figure.}
}
\value{
Invisibly returns the ggplot object.
}
\description{
Save a QQ plot of GWAS p-values
}
