% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{qqunif_plot}
\alias{qqunif_plot}
\title{QQ plot for p-values}
\usage{
qqunif_plot(
  pvalues,
  should.thin = T,
  thin.obs.places = 2,
  thin.exp.places = 2,
  xlab = expression(paste("Expected (", -log[10], " p-value)")),
  ylab = expression(paste("Observed (", -log[10], " p-value)")),
  draw.conf = TRUE,
  conf.points = 1000,
  conf.col = "lightgray",
  conf.alpha = 0.05,
  already.transformed = FALSE,
  pch = 20,
  aspect = "iso",
  prepanel = prepanel.qqunif,
  par.settings = list(superpose.symbol = list(pch = pch)),
  ...
)
}
\arguments{
\item{pvalues}{A numeric vector of p-values or a list of numeric vectors of p-values.}

\item{should.thin}{Logical indicating whether to thin the points for plotting. Default is TRUE.}

\item{thin.obs.places}{Number of decimal places to round the observed p-values for thinning. Default is 2.}

\item{thin.exp.places}{Number of decimal places to round the expected p-values for thinning. Default is 2.}
}
\description{
QQ plot for p-values
}
