% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{query_ot_api_variants}
\alias{query_ot_api_variants}
\title{Query Open Targets Platform API for variant information}
\usage{
query_ot_api_variants(variant_id = "19_44908822_C_T")
}
\arguments{
\item{variant_id}{A string representing the variant ID (e.g., "19_44908822_C_T").}
}
\value{
A data frame containing variant information.
}
\description{
Query Open Targets Platform API for variant information
}
