% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{query_vep_batch}
\alias{query_vep_batch}
\title{Query VEP API for a batch of variants}
\usage{
query_vep_batch(
  variant_batch,
  flag_pick = FALSE,
  include_hgvs = TRUE,
  include_domains = FALSE,
  include_regulatory = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{variant_batch}{Character vector of VCF-formatted variants}

\item{flag_pick}{Include flag_pick to mark the selected transcript}

\item{include_hgvs}{Include HGVS nomenclature}

\item{include_domains}{Include protein domains}

\item{include_regulatory}{Include regulatory consequences}
}
\value{
data.frame with VEP results
}
\description{
Query VEP API for a batch of variants
}
\keyword{internal}
