% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{reformat_names}
\alias{reformat_names}
\title{Rename columns in the dataset to standard names}
\usage{
reformat_names(ds, format)
}
\arguments{
\item{ds}{A data frame or tibble or duckdb tbl containing the GWAS summary statistics.}

\item{format}{The format of the input dataset (either "saige" or "regenie").}
}
\value{
An object of the same class as `ds` with renamed columns.
}
\description{
Rename columns in the dataset to standard names
}
