% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{reformat_summary_statistics}
\alias{reformat_summary_statistics}
\title{Reformat GWAS summary statistics from regenie or SAIGE}
\usage{
reformat_summary_statistics(file_path)
}
\arguments{
\item{file_path}{Path to the GWAS summary statistics file (parquet or CSV).}

\item{additional_cols}{A character vector of additional columns to keep from the original file.
Defaults to NULL.}
}
\value{
A data frame with the reformatted GWAS summary statistics. The standard
        columns are: CHR, POS, REF, ALT, P, BETA, SE.
}
\description{
This function takes a file path to a parquet or CSV file of GWAS summary statistics
output from regenie or SAIGE and reformats it to have a standard set of columns.
}
