% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{select_top_hits}
\alias{select_top_hits}
\alias{select_top_hits.GWASFormatter}
\alias{select_top_hits.data.frame}
\title{Pull top hits from a GWASFormatter or a data.frame/tibble}
\usage{
select_top_hits(x, threshold = 5e-08, ...)

\method{select_top_hits}{GWASFormatter}(x, threshold = 5e-08, ...)

\method{select_top_hits}{data.frame}(x, threshold = 5e-08, ...)
}
\arguments{
\item{x}{A GWASFormatter object, data.frame, or tibble.}

\item{threshold}{The p-value threshold to filter the top hits. Default is 5e-8.}

\item{...}{Additional arguments (unused).}
}
\value{
For GWASFormatter, a tibble of filtered hits; for data.frame/tibble, a filtered data.frame/tibble.
}
\description{
Pull top hits from a GWASFormatter or a data.frame/tibble
}
\section{Methods (by class)}{
\itemize{
\item \code{select_top_hits(GWASFormatter)}: Method for GWASFormatter objects

\item \code{select_top_hits(data.frame)}: Method for data.frame/tibble objects

}}
