% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_with_centromere}
\alias{annotate_with_centromere}
\title{Annotate top hits with centromere information}
\usage{
annotate_with_centromere(top_hits)
}
\arguments{
\item{top_hits}{A data frame containing the top hits.}
}
\value{
A data frame with the centromere information.
}
\description{
Annotate top hits with centromere information
}
