% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_with_chip_genes}
\alias{annotate_with_chip_genes}
\title{Annotate data with CHIP gene information}
\usage{
annotate_with_chip_genes(x, ...)
}
\description{
Annotate data with CHIP gene information
}
