% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{find_nearest_gene}
\alias{find_nearest_gene}
\title{Find the nearest gene for each top hit}
\usage{
find_nearest_gene(top_hits, threshold = 1e+05)
}
\arguments{
\item{top_hits}{A data frame containing the top hits.}

\item{threshold}{The distance threshold to consider a gene as nearest. Default is 1e5.}
}
\value{
A data frame with the nearest gene information.
}
\description{
Find the nearest gene for each top hit
}
