% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwasplot-package.R
\docType{package}
\name{gwasplot-package}
\alias{gwasplot}
\alias{gwasplot-package}
\title{gwasplot: High Performance GWAS Plotting And Annotation}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/weinstocklab/gwasplot}
  \item \url{https://weinstocklab.github.io/gwasplot/}
  \item Report bugs at \url{https://github.com/weinstocklab/gwasplot/issues}
}

}
\author{
\strong{Maintainer}: Josh Weinstock \email{josh.weinstock@emory.edu}

}
\keyword{internal}
