% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan}
\alias{manhattan}
\title{Plot a Manhattan plot from a gwas object}
\usage{
manhattan(gwas, output_prefix, lower_logp_threshold = 3)
}
\arguments{
\item{gwas}{A gwas object containing the data to plot.}

\item{output_prefix}{The prefix for the output file name.}

\item{lower_logp_threshold}{The lower threshold for the -log10(p-value) to plot. Default is 3.0.}
}
\description{
Plot a Manhattan plot from a gwas object
}
