% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{qqplot}
\alias{qqplot}
\title{QQ plot for p-values}
\usage{
qqplot(
  pvalues,
  should.thin = TRUE,
  thin.obs.places = 2,
  thin.exp.places = 2,
  xlab = expression(paste("Expected (", -log[10], " p-value)")),
  ylab = expression(paste("Observed (", -log[10], " p-value)")),
  draw.conf = TRUE,
  conf.points = 1000,
  conf.col = "lightgray",
  conf.alpha = 0.05,
  already.transformed = FALSE,
  point.size = 1.5,
  point.alpha = 0.7,
  point.color = "#3366FF",
  ...
)
}
\arguments{
\item{pvalues}{A numeric vector of p-values or a list of numeric vectors of p-values.}

\item{should.thin}{Logical indicating whether to thin the points for plotting. Default is TRUE.}

\item{thin.obs.places}{Number of decimal places to round the observed p-values for thinning. Default is 2.}

\item{thin.exp.places}{Number of decimal places to round the expected p-values for thinning. Default is 2.}
}
\description{
QQ plot for p-values
}
